# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Buty resource api",
        "description": "Buty Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
        "/api/users/{id}/calendar": {
            "get": {
                "tags": [
                    "User",
                    "Calendar"
                ],
                "summary": "Get user calendar",
                "description": "Returns user calendar",
                "operationId": "userCal",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "403": {
                        "description": "This action is unauthorized"
                    },
                    "404": {
                        "description": "Resource not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/groups/{id}/calendar": {
            "get": {
                "tags": [
                    "Group",
                    "Calendar"
                ],
                "summary": "Get Resource information",
                "description": "Returns group calendar",
                "operationId": "userCal",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "group id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "403": {
                        "description": "This action is unauthorized"
                    },
                    "404": {
                        "description": "Resource not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/{resource}/{id}": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "Get Resource information",
                "description": "Returns resource",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "resource",
                        "in": "path",
                        "description": "Resource name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "departments",
                                "formations"
                            ]
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Resource id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "403": {
                        "description": "This action is unauthorized"
                    },
                    "404": {
                        "description": "Resource not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        }
    }
}